Moved CommunicatingAgent classes to agents.py.
